Added derivation of core migrations and support for postgis types

+++ web/core_structures/migrations/00000000000005_create_samples_table/up.sql
@@ -9,7 +9,7 @@ CREATE TABLE IF NOT EXISTS samples (
     created_at TIMESTAMP NOT NULL DEFAULT NOW(),
     updated_by INTEGER NOT NULL,
     updated_at TIMESTAMP NOT NULL DEFAULT NOW(),
-    state_id INTEGER NOT NULL DEFAULT 1,
+    state_id SMALLINT NOT NULL DEFAULT 1,
     FOREIGN KEY (container_id) REFERENCES sample_containers(id) ON DELETE CASCADE,
     FOREIGN KEY (project_id) REFERENCES projects(id) ON DELETE CASCADE,
     FOREIGN KEY (created_by) REFERENCES users(id) ON DELETE CASCADE,
